requirements.txt
aiohttp==3.7.4
asgiref==3.3.1
async-timeout==3.0.1
attrs==20.3.0
backoff==1.10.0
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.5
chardet==3.0.4
click==7.1.2
cryptography==3.4.6
cycler==0.10.0
Django==3.1.7
django-request==1.5.6
Flask==1.1.2
google-api-core==1.26.0
google-api-python-client==1.12.8
google-auth==1.27.0
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.2
googleapis-common-protos==1.53.0
gspread==3.7.0
gunicorn==20.0.4
httplib2==0.19.0
idna==2.10
inline==0.0.1
itsdangerous==1.1.0
Jinja2==2.11.3
joblib==1.0.1
kiwisolver==1.3.1
MarkupSafe==1.1.1
matplotlib==3.3.4
multidict==5.1.0
numpy==1.20.1
oauth2client==4.1.3
oauthlib==3.1.0
packaging==20.9
pandas==1.2.2
Pillow==8.1.1
plotly==4.14.3
protobuf==3.15.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyOpenSSL==20.0.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
retrying==1.3.3
rsa==4.7.2
scikit-learn==0.24.1
scipy==1.6.1
seaborn==0.11.1
six==1.15.0
spotify==0.10.2
spotipy==2.17.1
sqlparse==0.4.1
threadpoolctl==2.1.0
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.3
Werkzeug==1.0.1
yarl==1.6.3
